Transaction Table updated

+++ Store/TTD_DataUsage.sql
@@ -2,6 +2,6 @@
 (
 	Id BIGINT IDENTITY(1,1) NOT NULL PRIMARY KEY, 
 	Month NVARCHAR(100) NOT NULL,	
-	Microsoft BIGINT NOT NULL,
-	Zoom BIGINT NOT NULL,
+	Microsoft Decimal NOT NULL,
+	Zoom Decimal NOT NULL,
 )
