+++ Store/TTD_SubscriptionSummary.sql
@@ -3,6 +3,6 @@
 	Id BIGINT IDENTITY(1,1) NOT NULL PRIMARY KEY,
 	Vendor	NVARCHAR(250) NOT NULL,	
 	Subscription NVARCHAR(250) NOT NULL,	
-	Total BIGINT NOT NULL,
-	AverageSpend BIGINT NOT NULL
+	Total Decimal NOT NULL,
+	AverageSpend Decimal NOT NULL
 )
